AppliedPython/ht02_Telegram_calories_tracker_bot/requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
googletrans==4.0.2
